Addressing second round of comments
