+++ src/cryptonote_protocol/cryptonote_protocol_handler.h
@@ -36,7 +36,6 @@
 
 #include <boost/program_options/variables_map.hpp>
 #include <string>
-#include <thread>
 
 #include "byte_slice.h"
 #include "math_helper.h"
@@ -210,7 +209,7 @@ namespace cryptonote
     epee::math_helper::once_a_time_seconds<43> m_bad_peer_checker;
     epee::math_helper::once_a_time_seconds<300> m_txpool_complement_checker;
     std::atomic<uint64_t> m_interval_peer_request_checker{P2P_DEFAULT_REQUEST_TIMEOUT}; // microseconds
-    std::function<uint64_t()> m_peer_request_interval; // set in ctor
+    const std::function<uint64_t()> m_peer_request_interval; // set in ctor
     epee::math_helper::once_a_time<decltype(m_peer_request_interval)> m_peer_request_checker;
     std::unordered_map<epee::net_utils::zone, unsigned int> m_max_out_peers;
     mutable epee::critical_section m_max_out_peers_lock;
