+++ src/cryptonote_protocol/request_manager.h
@@ -44,12 +44,13 @@
 #include <boost/uuid/uuid.hpp>
 #include <boost/uuid/uuid_io.hpp>
 #include <unordered_map>
+#include <shared_mutex>
 
 class request_manager {
 
 private:
   request_container m_requested_txs;
-  mutable epee::rw_mutex m_mutex;
+  mutable std::shared_timed_mutex m_mutex;
 
   request_manager(const request_manager &) = delete;
   request_manager &operator=(const request_manager &) = delete;
@@ -57,192 +58,31 @@ private:
   request_manager &operator=(request_manager &&) = delete;
 
 public:
-  request_manager() : m_requested_txs(), m_mutex() {}
+  request_manager();
 
-  bool remove_transaction(const crypto::hash &tx_hash) {
-    MINFO("Removing transaction: " << epee::string_tools::pod_to_hex(tx_hash));
-    epee::write_lock w_lock(m_mutex);
-    auto& requests_by_tx_hash = get_requests_by_tx_hash(m_requested_txs);
-    auto range = requests_by_tx_hash.equal_range(tx_hash);
-    if (range.first != range.second) {
-      size_t removed = std::distance(range.first, range.second);
-      requests_by_tx_hash.erase(range.first, range.second);
-      MINFO("Removed " << removed << " requests for tx " << epee::string_tools::pod_to_hex(tx_hash));
-      return true;
-    }
-    return false;
-  }
+  bool remove_transaction(const crypto::hash &tx_hash);
 
-  bool remove_requested_transaction(const boost::uuids::uuid &peer_id, const crypto::hash &tx_hash) {
-    MINFO("Removing transaction: " << epee::string_tools::pod_to_hex(tx_hash)
-                                    << " for peer: " << epee::string_tools::pod_to_hex(peer_id));
-    epee::write_lock w_lock(m_mutex);
-    auto& requests_by_peer_and_tx = get_requests_by_peer_and_tx(m_requested_txs);
-    auto it = requests_by_peer_and_tx.find(boost::make_tuple(peer_id, tx_hash));
-    if (it != requests_by_peer_and_tx.end()) {
-      requests_by_peer_and_tx.erase(it);
-      MINFO("Removed request for tx " << epee::string_tools::pod_to_hex(tx_hash)
-                                      << " from peer " << epee::string_tools::pod_to_hex(peer_id));
-      return true;
-    }
-    return false;
-  }
+  bool remove_requested_transaction(const boost::uuids::uuid &peer_id, const crypto::hash &tx_hash);
 
-  void remove_peer_requests(const boost::uuids::uuid &peer_id) {
-    MINFO("Removing all requests for disconnected peer: " << epee::string_tools::pod_to_hex(peer_id));
-    epee::write_lock w_lock(m_mutex);
-    auto& by_peer = get_requests_by_peer_id(m_requested_txs);
-    auto peer_range = by_peer.equal_range(peer_id);
-    if (peer_range.first != peer_range.second) {
-      size_t removed = std::distance(peer_range.first, peer_range.second);
-      by_peer.erase(peer_range.first, peer_range.second);
-      MINFO("Removed " << removed << " requests for peer " << epee::string_tools::pod_to_hex(peer_id));
-    }
-  }
+  void remove_peer_requests(const boost::uuids::uuid &peer_id);
 
-  void cleanup_stale_requests(std::time_t max_age_seconds) {
-    std::time_t now = std::time(nullptr);
-    std::time_t cutoff_time = now - max_age_seconds;
+  void cleanup_stale_requests(std::time_t max_age_seconds);
 
-    epee::write_lock w_lock(m_mutex);
-    size_t removed = 0;
-
-    auto it = m_requested_txs.begin();
-    while (it != m_requested_txs.end()) {
-      if (it->firstseen_timestamp < cutoff_time) {
-        MINFO("Removing stale request for tx " << epee::string_tools::pod_to_hex(it->tx_hash) 
-              << " from peer " << epee::string_tools::pod_to_hex(it->peer_id)
-              << ", age: " << (now - it->firstseen_timestamp) << " seconds");
-        it = m_requested_txs.erase(it);
-        ++removed;
-      } else {
-        ++it;
-      }
-    }
-
-    if (removed > 0) {
-      MINFO("Cleaned up " << removed << " stale requests");
-    }
-    return;
-  }
-
-  bool already_requested_tx(const crypto::hash &tx_hash) const {
-    epee::read_lock r_lock(m_mutex);
-    auto& by_tx_hash = get_requests_by_tx_hash(m_requested_txs);
-    for (const auto& req : boost::make_iterator_range(by_tx_hash.equal_range(tx_hash))) {
-      if (req.is_in_flight()) {
-        return true;
-      }
-    }
-    return false;
-  }
+  bool already_requested_tx(const crypto::hash &tx_hash) const;
 
   bool add_transaction(const crypto::hash &tx_hash,
-                       const boost::uuids::uuid &id, std::time_t first_seen) {
-    MINFO("Adding peer: " << epee::string_tools::pod_to_hex(id)
-                          << " to transaction: "
-                          << epee::string_tools::pod_to_hex(tx_hash)
-                          << ", first seen: " << first_seen);
-    epee::write_lock w_lock(m_mutex);
-    auto& by_peer_and_tx = get_requests_by_peer_and_tx(m_requested_txs);
-    auto it = by_peer_and_tx.find(boost::make_tuple(id, tx_hash));
-    if (it != by_peer_and_tx.end()) {
-      // already have this peer for this tx
-      MDEBUG("Peer " << epee::string_tools::pod_to_hex(id)
-                     << " already in request queue for tx "
-                     << epee::string_tools::pod_to_hex(tx_hash));
-      return false;
-
-    }
-    m_requested_txs.insert(request(id, tx_hash, first_seen));
-    return true;
-  }
+                       const boost::uuids::uuid &id, std::time_t first_seen);
 
   // Get the peer ID of the current in-flight request for a transaction, if any
   // true: found, false: not found or none in-flight
-  bool get_current_request_peer_id(const crypto::hash &tx_hash, boost::uuids::uuid &out) const {
-    epee::read_lock r_lock(m_mutex);
-    auto& by_tx_hash = get_requests_by_tx_hash(m_requested_txs);
-    auto it = by_tx_hash.find(tx_hash);
-    if (it == by_tx_hash.end()) {
-      MDEBUG("No requests for tx " << epee::string_tools::pod_to_hex(tx_hash));
-      return false;
-    }
-    for (const auto& req : boost::make_iterator_range(by_tx_hash.equal_range(tx_hash))) {
-      if (req.is_in_flight()) {
-        out = req.peer_id;
-        return true;
-      }
-    }
-    return false;
-  }
+  bool get_current_request_peer_id(const crypto::hash &tx_hash, boost::uuids::uuid &out) const;
 
   // Get next peer ID to request a transaction from, and mark it in-flight,
   // Remove the the current in-flight requests
-  boost::uuids::uuid request_from_next_peer(const crypto::hash &tx_hash, std::time_t now) {
-    epee::write_lock w_lock(m_mutex);
-    auto& by_tx_hash = get_requests_by_tx_hash(m_requested_txs);
-    auto it = by_tx_hash.find(tx_hash);
-    if (it == by_tx_hash.end()) {
-      MDEBUG("No requests for tx " << epee::string_tools::pod_to_hex(tx_hash));
-      return boost::uuids::nil_uuid();
-    }
-
-    // remove in flight request that have timed out
-    auto range = by_tx_hash.equal_range(tx_hash);
-    it = range.first;
-    while (it != range.second) {
-      if (it->is_in_flight()) {
-        MINFO("Erasing " << epee::string_tools::pod_to_hex(tx_hash)
-                               << " from peer "
-                               << epee::string_tools::pod_to_hex(it->peer_id));
-        it = by_tx_hash.erase(it);
-      } else {
-        ++it;
-      }
-    }
-
-    // Recompute range after potential erasures.
-    by_tx_hash = get_requests_by_tx_hash(m_requested_txs);
-    if (by_tx_hash.empty()) {
-      MINFO("No available requests for tx " << epee::string_tools::pod_to_hex(tx_hash));
-      return boost::uuids::nil_uuid();
-    }
-
-    // find the earliest non in-flight request
-    auto earliest_request = by_tx_hash.end();
-    it = by_tx_hash.find(tx_hash);
-    while (it != by_tx_hash.end() && it->tx_hash == tx_hash) {
-      if (!it->is_in_flight()) {
-        if (earliest_request == by_tx_hash.end() ||
-            it->firstseen_timestamp < earliest_request->firstseen_timestamp) {
-          earliest_request = it;
-        }
-      }
-      ++it;
-    }
-
-    if (earliest_request != by_tx_hash.end()) {
-      earliest_request->fly();
-      MINFO("Requesting tx " << epee::string_tools::pod_to_hex(tx_hash)
-                             << " from peer "
-                             << epee::string_tools::pod_to_hex(earliest_request->peer_id)
-                             << ", first seen " << earliest_request->firstseen_timestamp
-                             << ", now " << now);
-      return earliest_request->peer_id;
-    }
-
-    MINFO("No available requests for tx " << epee::string_tools::pod_to_hex(tx_hash));
-    return boost::uuids::nil_uuid();
-  }
+  boost::uuids::uuid request_from_next_peer(const crypto::hash &tx_hash, std::time_t now);
 
   void for_each_request(std::function<void(request_manager&, const request &, const std::time_t)> &f,
-                                            const std::time_t request_deadline) {
-    MINFO("Iterating over requested transactions");
-    for (const auto &request : m_requested_txs) {
-      f(*this, request, request_deadline);
-    }
-  }
+                        const std::time_t request_deadline);
 };
 
 #endif // CRYPTONOTE_PROTOCOL_REQUEST_MANAGER_H
